id_of function can return latest ID
Previous `id_of` returns the first selection of a list of IDs associated with a service name. This list of IDs exists because there may be naming colisons. This commit adds support for a new flag, `--newest`, which allows `id_of` to return the latest ID instead of just the first one.